+++ test/framework/src/main/java/org/elasticsearch/index/engine/EngineTestCase.java
@@ -498,6 +498,8 @@ public abstract class EngineTestCase extends ESTestCase {
      * Exposes a translog associated with the given engine for testing purpose.
      */
     public static Translog getTranslog(Engine engine) {
-        return engine.getTranslog();
+        assert engine instanceof InternalEngine : "only InternalEngines have translogs, got: " + engine.getClass();
+        InternalEngine internalEngine = (InternalEngine) engine;
+        return internalEngine.getTranslog();
     }
 }
